Refactor lib into class-specific files
